requirements.txt
django-cms<3.3
django-treebeard>=2.0

https://github.com/divio/djangocms-column/archive/master.zip?1449652964.68#egg=djangocms-column
https://github.com/divio/djangocms-flash/archive/master.zip?1449652964.68#egg=djangocms-flash
https://github.com/divio/djangocms-googlemap/archive/master.zip?1449652964.68#egg=djangocms-googlemap
https://github.com/divio/djangocms-inherit/archive/master.zip?1449652964.68#egg=djangocms-inherit
https://github.com/divio/djangocms-link/archive/master.zip?1449652964.68#egg=djangocms-link
https://github.com/divio/djangocms-style/archive/master.zip?1449652964.68#egg=djangocms-style
easy_thumbnails
django-filer>=0.9.12
cmsplugin-filer>=0.10.2
djangocms-admin-style>=1.0.6
djangocms-text-ckeditor>=2.8.0
Django<1.9
pytz
django-reversion>=1.8.7,<1.9
django-classy-tags>=0.3.4.1
html5lib
Pillow>=2.3
django-sekizai>=0.7
six

# additional requirements
dj-database-url==0.3.0
whitenoise==2.0.6
gunicorn==19.4.1
psycopg2==2.6.1
django-laterpay==1.2.0
